Do not require pandas to run unit tests
